Refactor Automovil class for better structure
Updated the Automovil class to fix method return types and improve code structure.

+++ actividad02/ejercicio03/Automovil.java
@@ -119,7 +119,7 @@ public class Automovil {
 
     }
 
-    int tiempoLlegada(int distancia) {
+    double tiempoLlegada(int distancia) {
         if (velocidadActual > 0) {
             return distancia / velocidadActual;
         } else {
@@ -145,3 +145,4 @@ public class Automovil {
         System.out.println("Color: " + color + "\n");
     }   
 }
+
